update models, include discount

+++ db/migrate/20210807181511_create_discounts.rb
@@ -2,7 +2,7 @@ class CreateDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :discounts do |t|
       t.integer :quantity
-      t.integer :discount
+      t.float :discount
       t.references :merchant, foreign_key: true
     end
   end
